+++ material/src/material.ts
@@ -1,5 +1,5 @@
 export * from './modules/dialogMetadataSelector.module';
 export * from './components/verticalDragNDropSelection/types';
 export * from './components/verticalDragNDropSelection';
-export {DialogMetadataSelectorComponent as DialogMetadataSelectorNgComponent} from './plugins/metadataSelector/components';
+export * from './plugins/metadataSelector/components';
 export * from './plugins/metadataSelector';